requirement.txt
bcrypt==3.2.0
cffi==1.14.3
cryptography==3.2.1
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
MarkupSafe==1.1.1
netmiko==3.3.2
ntc-templates==1.6.0
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
schedule==0.6.0
scp==0.13.3
six==1.15.0
tenacity==6.2.0
textfsm==1.1.0
xlrd==1.2.0
XlsxWriter==1.3.7
